refactor bootstrap demos
1. add .gitignore file; 2. refactoring CertificateUtil; 3. moving bootstrap demos into a single package; 4. optimize bootstrap demos
